+++ src/amuse/community/bhtree/tests/conftest.py
@@ -2,41 +2,15 @@ from pytest import fixture
 from amuse.community.bhtree.interface import BHTree
 
 from amuse.units import nbody_system
-from amuse.support import literature
-literature.TrackLiteratureReferences.suppress_output()
 
 
 from amuse.units import nbody_system
 from amuse.units import units
 from amuse.units import quantities
 
-# Factory to create bhtrees. Handles teardown for all
-# Follows https://docs.pytest.org/en/stable/how-to/fixtures.html#factories-as-fixtures
 @fixture()
-def make_nbody_instance():
-    created_instances = []
+def nbody_implementation():
+    return BHTree
 
-    def _make_instance(*args, **kwargs):
-        instance = BHTree(*args, **kwargs)
-        created_instances.append(instance)
-        return instance
-
-    yield _make_instance
-
-    for instance in created_instances:
-        instance.cleanup_code()
-        instance.stop()
-
-@fixture()
-def nbody_instance(make_nbody_instance):
-    yield make_nbody_instance()
-
-
-@fixture()
-def nbody_instance_kg(make_nbody_instance): # for test4, test6, test7, test11, test12, test13
-    convert_nbody = nbody_system.nbody_to_si(5.0 | units.kg, 10.0 | units.m)
-    instance = make_nbody_instance(convert_nbody)
-    instance.commit_parameters()
-    yield instance
